cast course duration

+++ courses/models.py
@@ -1,6 +1,6 @@
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models, transaction
-from django.db.models import Q, Sum, FloatField
+from django.db.models import Q, Sum, DecimalField
 from django.db.models.functions import Cast
 
 from django.conf import settings
@@ -96,7 +96,7 @@ class Course(models.Model):
         return self.content.count()
 
     def get_duration(self):
-        duration = self.content.annotate(sum=Cast('duration', output_field=FloatField())).aggregate(sum=Sum('duration'))['sum']
+        duration = self.content.annotate(sum=Cast('duration', output_field=DecimalField())).aggregate(sum=Sum('duration'))['sum']
         return duration
